Fixes to the thin_by_cell_time function
I have added two fixes to the thin_by_cell_time function.

First, the step where the data was subsetted by time_col in base R and then in tidyverse style threw errors because the base R subsetting removed the name of the selected column. I fixed this by changing it to subset tidyverse-style only.

Second, the documentation on the function was insufficient and caused confusion. The code requires the input SpatRaster to have time set as POSIXt for all layers or the time_indices step throws an error. I have added to the documentation to make it clear that terra::time must be used to set SpatRaster layers to POSIXt.