Bug fix to coordinate columns in thin_by_cell
I was having problems with a sf data object that already had its coordinates stashed in X, Y columns, so I added some code to check if the correct coordinates are already there. If they are, coordinates are not bound. If there are X and Y columns that don't match the coordinates, they are replaced with a warning.