explain_tidysdm.R a bug in explain_tidysdm.linear_stack
Using explain_tidusdm for a linear_stack did not work well, receiving an error "Error in model_info.linear_stack(model, is_multiclass = task_subtype) : tidysdm linear_stack can not be multiclass"

This might be because DALEX considers my linear_stack a multiclass because it contains both presence and background as two levels of the outcome factor. Adjustung the predict_function to only select .pred_presence resolved this problem, and the explainer for linear_stack now works for both var importance plot and pdb plot.

It still does not work for individual plots, which probably makes sence. Or would we like to produce plots for individual models from which the stack consists despite not including weights?